requirements.txt
streamlit
pandas
numpy
PyPDF2
faiss-cpu
tiktoken
openai